new_requirements.txt
absl-py==2.3.1
certifi==2026.1.4
cffi==1.17.1
charset-normalizer==3.4.4
contourpy==1.1.1
cryptography==46.0.4
cycler==0.12.1
etils==1.3.0
filelock==3.16.1
fonttools==4.57.0
fsspec==2025.3.0
glfw==2.10.0
google-auth==2.48.0
google-auth-oauthlib==1.0.0
grpcio==1.70.0
idna==3.11
imageio==2.35.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
# Editable install with no version control (isaacgym==1.0rc4)
-e /home/sviswasam/dr/isaacgym/python
Jinja2==3.1.6
kiwisolver==1.4.7
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
mujoco==3.2.3
networkx==3.1
ninja==1.13.0
numpy==1.20.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
oauthlib==3.3.1
packaging==26.0
pillow==10.4.0
protobuf==5.29.6
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.23
PyOpenGL==3.1.10
pyparsing==3.1.4
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
-e git+https://github.com/leggedrobotics/rsl_rl.git@2ad79cf0caa85b91721abfe358105f869a784121#egg=rsl_rl
scipy==1.10.1
six==1.17.0
sympy==1.13.3
tensorboard==2.14.0
tensorboard-data-server==0.7.2
torch==2.3.1+cu121
torchaudio==2.3.1+cu121
torchvision==0.18.1+cu121
triton==2.3.1
typing_extensions==4.13.2
-e git+https://github.com/unitreerobotics/unitree_rl_gym.git@276801e46c5d433564f24658bac64f254b7d2d4b#egg=unitree_rl_gym
urllib3==2.2.3
Werkzeug==3.0.6
zipp==3.20.2
